modify proto file path to support apollo 8.0

+++ src/carla_cyber_bridge/radar.py
@@ -14,7 +14,7 @@ import numpy as np
 
 from carla_cyber_bridge.sensor import Sensor, create_cloud
 
-from modules.drivers.proto.pointcloud_pb2 import PointXYZIT, PointCloud
+from modules.common_msgs.sensor_msgs.pointcloud_pb2 import PointXYZIT, PointCloud
 
 
 class Radar(Sensor):
@@ -77,7 +77,7 @@ class Radar(Sensor):
             point.x = detection.depth * np.cos(detection.azimuth) * np.cos(-detection.altitude)
             point.y = detection.depth * np.sin(-detection.azimuth) * np.cos(detection.altitude)
             point.z = detection.depth * np.sin(detection.altitude)
-            point.intensity = long(detection.depth)
+            point.intensity = int(detection.depth)
             points.append(point)
 
         radar_msg = create_cloud(self.get_msg_header(
